Minor improvements

+++ darkpay-auto-updater.sh
@@ -10,16 +10,17 @@ if [ "$latestVersion" = "$currentVersionShort"  -o "$latestVersion" = "$currentV
     echo "Already at latest version." #> /dev/null
 else
    cd /root
-   service darkpay stop
    fullpath=$(echo $json | jq .clients | jq .darkpayd | jq .platforms | jq .linux | jq .x64 | jq .download | jq .ur$
    bin=$(echo $json | jq .clients | jq .darkpayd | jq .platforms | jq .linux | jq .x64 | jq .download | jq .bin | t$
    filename=$(echo $fullpath | rev | cut -d \/ -f 1 | rev)
    folder=$(echo $bin | cut -d \/ -f 1)
    wget $fullpath
-   tar -xzvf $filename
-   rm /usr/local/bin/darkpayd
-   cp $bin /usr/local/bin/
-   service darkpay start
-   rm -R $folder
-   rm $filename
+   if [ -e $filename ]; then
+      tar -xzvf $filename
+      service darkpay stop && rm /usr/local/bin/darkpayd
+      cp $bin /usr/local/bin/
+      service darkpay start
+      rm -R $folder
+      rm $filename
+    fi
 fi
